Move new test to test_core
